+++ packages/serverless-api/src/index.ts
@@ -1,3 +1,9 @@
+/**
+ * @module @twilio-labs/serverless-api
+ * @preferred
+ *
+ * Main entry point of the project
+ */
 export * from './api';
 export * from './client';
 export * from './types';
