+++ packages/serverless-api/src/types/client.ts
@@ -1,0 +1,13 @@
+/**
+ * Config to set up a API request client
+ */
+export type ClientConfig = {
+  /**
+   * Twilio AccountSID or API Key
+   */
+  accountSid: string;
+  /**
+   * Twilio Auth Token or API Secret
+   */
+  authToken: string;
+};
