+++ packages/serverless-api/src/types/index.ts
@@ -1,0 +1,6 @@
+export * from './activate';
+export * from './client';
+export * from './deploy';
+export * from './generic';
+export * from './list';
+export * from './serverless-api';
